IN-212: refactor: created hook for functions of creating comment
